requirements.txt
PyQt5
PyOpenGL
PyOpenGL_accelerate
PyGLM